+++ SMB3_C++/SMB3_C++/assets/headers/HitBox.h
@@ -13,8 +13,8 @@ private:
 public:
 	RECTF GetBoundingBox(int = 0) const;
 
-	unsigned int GetWidth(int = 0) const;
-	unsigned int GetHeight(int = 0) const;
+	float GetWidth(int = 0) const;
+	float GetHeight(int = 0) const;
 
 	void AddHitBox(RECTF);
 };
