+++ SMB3_C++/SMB3_C++/assets/source/HitBox.cpp
@@ -4,11 +4,11 @@ RECTF HitBox::GetBoundingBox(int id) const {
 	return hitboxes.at(id);
 }
 
-unsigned int HitBox::GetWidth(int id) const {
+float HitBox::GetWidth(int id) const {
 	return hitboxes.at(id).right;
 }
 
-unsigned int HitBox::GetHeight(int id) const {
+float HitBox::GetHeight(int id) const {
 	return hitboxes.at(id).bottom;
 }
